requirements.txt
requests>=2.13.0
MySQL-python>=1.2.3
mysql-replication>=0.11